+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog for Bearded Theme
 
+## 8.3.2 - 2023-07-28
+
+### Enhancements
+
+- Improve readibility or peekview
+
+### Fixes
+
+- Fix sticky scroll background on peekview (Thanks [aiday-mar](https://github.com/BeardedBear/bearded-theme/issues/127))
+
 ## 8.3.1 - 2023-07-26
 
 - Fix drop color on activity bar buttons (Thanks [Loskir](https://github.com/BeardedBear/bearded-theme/issues/125))